+++ src/index.ts
@@ -38,10 +38,7 @@ export * from './type/patterns/index'
 export * from './type/registry/index'
 export * from './type/sets/index'
 export * from './type/symbols/index'
-// ------------------------------------------------------------------
-// Parse
-// ------------------------------------------------------------------
-export * from './parse/index'
+
 // ------------------------------------------------------------------
 // Types
 // ------------------------------------------------------------------
