+++ src/syntax/parsebox/runtime/token.ts
