+++ src/syntax/parsebox/static/token.ts
