+++ test/runtime/index.ts
@@ -8,7 +8,7 @@ TypeSystemPolicy.InstanceMode = 'freeze'
 import './compiler/index'
 import './compiler-ajv/index'
 import './errors/index'
-import './parse/index'
+import './syntax/index'
 import './system/index'
 import './type/index'
 import './value/index'
