--- test/runtime/parse/index.ts
@@ -1,1 +1,0 @@
-import './parse'
