+++ test/runtime/syntax/index.ts
@@ -1,0 +1,1 @@
+import './syntax'
